requirements.txt
Babel==1.3
dj-database-url==0.3.0
Django==1.8.4
django-appconf==1.0.1
django-apps==0.1.0
django-compressor==1.5
django-debug-toolbar==1.3.2
django-debug-toolbar-template-timings==0.6.4
django-extensions==1.5.6
django-extra-views==0.6.4
django-haystack==2.3.1
django-newsletter-subscription==0.2.0
django-oscar==1.1.1
django-postgrespool==0.3.0
django-rosetta==0.7.6
django-tables2==1.0.4
django-treebeard==3.0
django-widget-tweaks==1.3
django-yandex-money==1.1.2
factory-boy==2.4.1
gunicorn==19.3.0
lxml==3.4.4
microsofttranslator==0.5
mock==1.0.1
nltk==3.0.5
numpy==1.9.2
pandas==0.16.2
phonenumbers==6.3.0
Pillow==2.7.0
pluggy==0.3.1
polib==1.0.7
psycopg2==2.6.1
purl==1.1
py==1.4.30
pycountry==1.15
pysolr==3.3.2
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
rosetta==0.3
scipy==0.15.1
six==1.9.0
sorl-thumbnail==12.2
South==1.0.2
SQLAlchemy==1.0.8
sqlparse==0.1.16
tox==2.1.1
Unidecode==0.4.18
virtualenv==13.1.2
wheel==0.24.0
whitenoise==2.0.3
Whoosh==2.7.0
zope.interface==4.1.2
